Changed the type of `DvdEpisodeNumber` from `int?` to `decimal?` in the `EpisodeRecord` and `BasicEpisode` models

+++ src/TvDbSharper/Clients/Episodes/Json/EpisodeRecord.cs
@@ -25,7 +25,7 @@
 
         public string DvdDiscid { get; set; }
 
-        public int? DvdEpisodeNumber { get; set; }
+        public decimal? DvdEpisodeNumber { get; set; }
 
         public int? DvdSeason { get; set; }
